+++ feature/namecard/src/main/java/com/wepli/feature/photocard/detail/viewmodel/PhotoCardDetailViewModel.kt
@@ -4,7 +4,7 @@ import base.BaseMviViewModel
 import com.wepli.feature.photocard.detail.mvi.PhotoCardDetailEffect
 import com.wepli.feature.photocard.detail.mvi.PhotoCardDetailIntent
 import com.wepli.feature.photocard.detail.mvi.PhotoCardDetailUiState
-import com.wepli.shared.feature.uimodel.namecard.PhotoCardUiData
+import com.wepli.shared.feature.uimodel.photocard.PhotoCardUiData
 import com.wepli.uimodel.music.SongUiData
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
