+++ feature/namecard/src/main/java/com/wepli/feature/photocard/result/mvi/PhotoCardResultMvi.kt
@@ -3,7 +3,7 @@ package com.wepli.feature.photocard.result.mvi
 import base.Intent
 import base.SideEffect
 import base.UiState
-import com.wepli.shared.feature.uimodel.namecard.PhotoCardUiData
+import com.wepli.shared.feature.uimodel.photocard.PhotoCardUiData
 import com.wepli.shared.feature.uimodel.user.UserUiData
 
 data class PhotoCardResultUiState(
