+++ feature/namecard/src/main/java/com/wepli/feature/photocard/result/navigation/PhotoCardResultNavigation.kt
@@ -8,7 +8,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import com.wepli.feature.photocard.result.PhotoCardResultScreenRoute
 import com.wepli.navigator.feature.namecard.PhotoCardRoute
-import com.wepli.shared.feature.uimodel.namecard.PhotoCardUiData
+import com.wepli.shared.feature.uimodel.photocard.PhotoCardUiData
 import extensions.parseFromJson
 import extensions.toJsonString
